"Fixed" graphical issue with the SamplesTableViewHeader
When returning from the viro/react/cardboard views/scenes back into the native
app, the device is still sometimes in landscape orientation (even though my view
controller is set to never do so) and iOS seems to want to resize that header
view and set it to the width + height of the phone in landscape mode (it ignores
the frame I gave it earlier). So the way around this was to simply make the top
level view in the header transparent so when it resizes after viewDidAppear in the
containing viewcontroller, it doesn't "look" like it was wrongly sized because the
view is invisible but the view in the header is perfectly sized.